git: update to v1.7.10
This pushes the git package to the recently released v1.7.10,
and rebases all of our patches on top of that. As usual,
diffs of diffs are hard to read, so here's a quick run-down
on what changed:

  1. The archive-gz-external, git-fetch-performance, and
     remove-duplicate-dot-have-lines patches are now
     obsolete and are gone.

  2. The git-stats patch can drop the unix-socket code which
     was previously backported to v1.7.5.4 from a more
     recent git.

  3. The pack-object-race series is mostly upstream now; it
     now contains just our special "do not even bother
     warning" patch.

Other than that, it was just a few slight conflict
resolutions and cleanups.