bin: add some simple scripts for maintaining patches
Working with raw patches is annoying. These scripts let you
extract the patches into a set of git branches, and then
convert them back. So to move patches in the git package
from v1.7.5.4 to v1.7.10, you could do something like:

  $ cd packages/git
  $ ../../bin/patches-to-repo /your/git/repo v1.7.5.4

  $ cd /your/git/repo
  $ git rebase v1.7.10 brew2deb/archive-gz-external
  $ git rebase v1.7.10 brew2deb/git-stats
  [...etc...]

  $ cd /your/brew2deb/packages/git
  $ ../../bin/repo-to-patches /your/git/repo v1.7.10

After which all of your patches would be updated, and
obsolete ones removed entirely. It's safe to run
repo-to-patches as many times as you like, so you can modify
a branch and re-run if you want to tweak it. You can also
add new branches in the brew2deb hierarchy, and
repo-to-patches will automatically bring them in as new
patches.