git: fix minor patch conflict
There is a textual conflict between git-stats and the
advertise-alternates patches. Let's move the git-status
patch to happen first, as it is the most likely to need
updating in the future, which means it's easier to work with
directly on top of v1.7.10. Then we can just tweak the
advertise-alternates patch to avoid the conflict it.